# openapi.yaml
...
openapi: 3.1.0
info:
  title: 大模型安全评测系统 API
  description: 大模型安全评测系统后端接口骨架。本服务默认监听本机，不包含登录、用户、角色权限和多租户。
  version: 1.0.0
paths:
...
  /api/v1/benchmarks/import:
    post:
      tags:
      - Benchmark
      summary: Import Benchmark
      operationId: importBenchmark
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_importBenchmark'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_dict_str__Any__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - LocalToken: []
  /api/v1/benchmarks:
    get:
      tags:
      - Benchmark
      summary: List Benchmarks
      operationId: listBenchmarks
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_Benchmark__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/benchmarks/{benchmark_id}/versions:
    get:
      tags:
      - Benchmark
      summary: List Benchmark Versions
      operationId: listBenchmarkVersions
      security:
      - LocalToken: []
      parameters:
      - name: benchmark_id
        in: path
        required: true
        schema:
          type: integer
          title: Benchmark Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_BenchmarkVersion__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/test-cases:
    get:
      tags:
      - 测试用例
      summary: List Test Cases
      operationId: listTestCases
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: benchmark_version_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Benchmark Version Id
      - name: risk_category_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Category Code
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_TestCase__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/test-cases/{test_case_id}:
    patch:
      tags:
      - 测试用例
      summary: Update Test Case
      operationId: updateTestCase
      security:
      - LocalToken: []
      parameters:
      - name: test_case_id
        in: path
        required: true
        schema:
          type: integer
          title: Test Case Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TestCase_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/models:
    get:
      tags:
      - 模型
      summary: List Models
      operationId: listModels
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: usage_scope
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Usage Scope
      - name: enabled
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_Model__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 模型
      summary: Create Model
      operationId: createModel
      security:
      - LocalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Model_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/files:
    get:
      tags:
      - 文件
      summary: List Files
      operationId: listFiles
      security:
      - LocalToken: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: owner_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Type
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Owner Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaginatedData_StoredFile__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 文件
      summary: Upload File
      operationId: uploadFile
      security:
      - LocalToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_uploadFile'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_StoredFile_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/{file_id}:
    get:
      tags:
      - 文件
      summary: Get File
      operationId: getFile
      security:
      - LocalToken: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: integer
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_StoredFile_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/{file_id}/download:
    get:
      tags:
      - 文件
      summary: Download File
      operationId: downloadFile
      security:
      - LocalToken: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: integer
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    ApiResponse_Model_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/Model'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[Model]
...
    ApiResponse_PaginatedData_BenchmarkVersion__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_BenchmarkVersion_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[BenchmarkVersion]]
    ApiResponse_PaginatedData_Benchmark__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_Benchmark_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[Benchmark]]
...
    ApiResponse_PaginatedData_Model__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_Model_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[Model]]
...
    ApiResponse_PaginatedData_StoredFile__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_StoredFile_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[StoredFile]]
...
    ApiResponse_PaginatedData_TestCase__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/PaginatedData_TestCase_'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[PaginatedData[TestCase]]
...
    ApiResponse_StoredFile_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/StoredFile'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[StoredFile]
...
    ApiResponse_TestCase_:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/TestCase'
          - type: 'null'
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[TestCase]
    ApiResponse_dict_str__Any__:
      properties:
        code:
          type: string
          title: Code
          default: OK
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Data
        request_id:
          type: string
          title: Request Id
          default: req_placeholder
          examples:
          - req_01J8Z6K9YQ6M2H
      type: object
      title: ApiResponse[dict[str, Any]]
...
    Body_importBenchmark:
      properties:
        name:
          type: string
          title: Name
        version:
          type: string
          title: Version
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        risk_taxonomy_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Risk Taxonomy Id
      type: object
      required:
      - name
      - version
      - file
      title: Body_importBenchmark
    Body_uploadFile:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        owner_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Type
        owner_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Owner Id
      type: object
      required:
      - file
      title: Body_uploadFile
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Model:
      properties:
        model_id:
          type: integer
          title: Model Id
        name:
          type: string
          title: Name
        provider:
          type: string
          title: Provider
        adapter_type:
          $ref: '#/components/schemas/ModelAdapter'
        base_url:
          type: string
          title: Base Url
        model_name:
          type: string
          title: Model Name
        secret_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Ref
        usage_scope:
          $ref: '#/components/schemas/ModelUsageScope'
        enabled:
          type: boolean
          title: Enabled
          default: true
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - model_id
      - name
      - provider
      - adapter_type
      - base_url
      - model_name
      - usage_scope
      title: Model
    ModelAdapter:
      type: string
      enum:
      - openai_compatible
      - ollama
      - custom_http
      - mock
      title: ModelAdapter
...
    ModelCreateRequest:
      properties:
        name:
          type: string
          title: Name
        provider:
          type: string
          title: Provider
        adapter_type:
          $ref: '#/components/schemas/ModelAdapter'
        base_url:
          type: string
          title: Base Url
        model_name:
          type: string
          title: Model Name
        secret_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Ref
        default_params:
          additionalProperties: true
          type: object
          title: Default Params
        usage_scope:
          $ref: '#/components/schemas/ModelUsageScope'
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - name
      - provider
      - adapter_type
      - base_url
      - model_name
      - usage_scope
      title: ModelCreateRequest
    ModelUsageScope:
      type: string
      enum:
      - target
      - judge
      - both
      title: ModelUsageScope
...
    PaginatedData_BenchmarkVersion_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/BenchmarkVersion'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[BenchmarkVersion]
    PaginatedData_Benchmark_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Benchmark'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[Benchmark]
...
    PaginatedData_Model_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Model'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[Model]
...
    PaginatedData_StoredFile_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/StoredFile'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[StoredFile]
...
    PaginatedData_TestCase_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TestCase'
          type: array
          title: Items
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
          default: 20
        total:
          type: integer
          minimum: 0.0
          title: Total
          default: 0
        has_next:
          type: boolean
          title: Has Next
          default: false
      type: object
      title: PaginatedData[TestCase]
...
    StoredFile:
      properties:
        file_id:
          type: integer
          title: File Id
        original_name:
          type: string
          title: Original Name
        mime_type:
          type: string
          title: Mime Type
        size_bytes:
          type: integer
          title: Size Bytes
        sha256:
          type: string
          title: Sha256
        owner_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Type
        owner_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Owner Id
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - file_id
      - original_name
      - mime_type
      - size_bytes
      - sha256
      title: StoredFile
...
    TestCase:
      properties:
        test_case_id:
          type: integer
          title: Test Case Id
        benchmark_version_id:
          type: integer
          title: Benchmark Version Id
        external_id:
          type: string
          title: External Id
        prompt:
          type: string
          title: Prompt
        system_prompt:
          anyOf:
          - type: string
          - type: 'null'
          title: System Prompt
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        source_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Label
        risk_categories:
          items:
            $ref: '#/components/schemas/RiskCategoryRef'
          type: array
          title: Risk Categories
        status:
          type: string
          title: Status
          default: active
      type: object
      required:
      - test_case_id
      - benchmark_version_id
      - external_id
      - prompt
      title: TestCase
    TestCaseUpdateRequest:
      properties:
        prompt:
          anyOf:
          - type: string
          - type: 'null'
          title: Prompt
        system_prompt:
          anyOf:
          - type: string
          - type: 'null'
          title: System Prompt
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        risk_category_ids:
          anyOf:
          - items:
              type: integer
            type: array
          - type: 'null'
          title: Risk Category Ids
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: TestCaseUpdateRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    LocalToken:
      type: apiKey
      in: header
      name: X-Local-Token